requirements.txt
Flask==3.1.0
numpy==2.2.3
Pillow==11.1.0
opencv-python==4.11.0.86
pytest==8.3.5
gunicorn==23.0.0